updated some names and worked on cache, need to read cache from another place
