+++ src/components/authorization/authorization.scss
@@ -68,6 +68,14 @@
   }
 }
 
+.home-link {
+  margin-bottom: 2rem;
+
+  a {
+    color: $teal-color;
+  }
+}
+
 .login-link {
   margin-top: 1rem;
