+++ src/components/authorization/registration/registration.spec.jsx
