+++ src/components/friend-item/friend-item.jsx
@@ -4,10 +4,10 @@ import { Link } from 'react-router-dom';
 import percentMapper from '../../helpers/percent-mapper';
 
 const FriendItem = ({ data, username }) => {
-  const { name, icon, friendId, friendScore } = data;
+  const { name, icon, id, friendScore } = data;
 
   return (
-    <Link to={`/${username}/friends/${friendId}`} className="friend-wrapper">
+    <Link to={`/${username}/friends/${id}`} className="friend-wrapper">
       <div className="friend-item">
         <p className="friend-name">{name}</p>
         <div className="icon-container">
@@ -27,7 +27,7 @@ FriendItem.propTypes = {
     name: PropTypes.string.isRequired,
     icon: PropTypes.string.isRequired,
     friendScore: PropTypes.number.isRequired,
-    friendId: PropTypes.string.isRequired,
+    id: PropTypes.string.isRequired,
   }),
   username: PropTypes.string,
 };
