+++ src/components/friend-items/friend-items.jsx
@@ -5,7 +5,7 @@ import { FriendItem } from '..';
 
 const FriendItems = ({ friends, username }) => (
   <div className="friend-items content-wrapper">
-    {friends.map((data) => <FriendItem data={data} key={data.friendId} username={username} />)}
+    {friends.map((data) => <FriendItem data={data} key={data.id} username={username} />)}
   </div>
 );
