+++ src/graphql/cache-queries/index.js
@@ -1,0 +1,1 @@
+export { default as LOGIN_USER_CACHE_QUERY } from './login-user-cache';
