+++ src/graphql/error-link.js
@@ -1,12 +1,14 @@
+/* eslint-disable no-console */
 import { onError } from 'apollo-link-error';
 
 const errorLink = onError(({ graphQLErrors, networkError }) => {
   if (graphQLErrors) {
-    graphQLErrors.map(({ message, locations, path }) =>
-      console.log(`[GraphQL error]: ${message}`) // eslint-disable-line
-      // console.log(`[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`)
-    );
+    graphQLErrors.map(({ message, locations, path, extensions }) => {
+      if (extensions.code === 'UNAUTHENTICATED') return console.log('Unauthorized');
+      return console.log(`[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`);
+    });
   }
+
   if (networkError) {
     if (networkError.statusCode === '401') {
       console.log('Error: Unauthorized!'); // eslint-disable-line
