+++ src/graphql/queries/index.js
@@ -1,3 +1,3 @@
-export { default as USER_QUERY } from './get-user-query.js';
-export { default as FRIEND_QUERY } from './get-friend-query.js';
-export { default as FRIENDS_QUERY } from './get-friends-query.js';
+export { default as USER_QUERY } from './get-user.js';
+export { default as FRIEND_QUERY } from './get-friend.js';
+export { default as FRIENDS_QUERY } from './get-friends.js';
